Revert "add Skills component"
