+++ package.json
@@ -13,18 +13,13 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.2.3",
-    "node-sass": "^8.0.0",
-    "progressbar.js": "^1.1.0",
     "react": "^18.2.0",
-    "react-animated-progress-bar": "^1.1.2",
     "react-dark-theme": "^1.1.1",
     "react-dom": "^18.2.0",
-    "react-progressbar.js": "^0.2.0",
     "react-rounded-image": "^2.0.14",
     "react-router": "^6.4.3",
     "react-router-dom": "^6.4.3",
     "react-scripts": "^2.1.3",
-    "react-skillbars": "^2.1.0",
     "reactstrap": "^9.1.5",
     "web-vitals": "^2.1.4"
   },
