+++ src/App.css
@@ -1,10 +1,7 @@
 .App {
   text-align: center;
-  
 }
-.navbar {
-  z-index: 10; /* or any other value higher than the z-index of .stack-list */
-}
+
 .App-logo {
   height: 40vmin;
   pointer-events: none;
