+++ src/App.js
@@ -10,38 +10,40 @@ import Community from './components/Community';
 import Contact from './components/Contact';
 import Donate from './components/Donate';
 import Projects from './components/Projects';
-import Navbar from "./components/Navbar/Navbar";
+import Navbar from "./components/Navbar";
 import Playlist from './components/Playlist';
 import DarkTheme, { createTheme } from 'react-dark-theme'
 
 const lightTheme = {
   background: 'white',
   text: 'black',
-  
 }
 
 const darkTheme = {
   background: "	#181818",
   text: 'white',
-
 }
 const myTheme = createTheme(darkTheme, lightTheme)
 export default function App() {
 
   return (
   <div> <DarkTheme light={lightTheme} dark={darkTheme} /> 
-    <Router>         
+
+    <Router>
+
       <Navbar />
-      <Routes>
+
+        <Routes>
           <Route path="/" element={<Home/>} />
           <Route path="/community" element={<Community/>} />
           <Route path="/contact" element={<Contact/>} />
           <Route path="/donate" element={<Donate/>} />
           <Route path="/projects" element={<Projects/>}/>
           <Route path='/playlist' element={<Playlist/>}/>
-      </Routes>
-    </Router>
-  </div>
+</Routes>
+       
+        </Router>
+        </div>
        
   );
