--- src/components/Skills/languages.css
@@ -1,25 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat&display=swap');
-.stacks h2 {
-    font-family: 'Montserrat';
-   text-align: center; }
-  
-  .stacks__wrap {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    align-items: center;
-    margin-top: 2em; }
-    .stacks__wrap .stack-list {
-z-index: 5;
-      margin: 1em 2em; }
-      .stacks__wrap .stack-list__name-wrap {
-        display: flex;
-        justify-content: center;
-        align-items: center; }
-        .stacks__wrap .stack-list__name-wrap img {
-          height: 25px; }
-        .stacks__wrap .stack-list__name-wrap p {
-          
-          font-weight: 400;
-          margin-left: 1em;
-         }
