+++ src/components/css/Home.css
@@ -1,0 +1,56 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
+
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+    font-size: 14px;
+}
+
+body {
+    background-color: #e2e8f0;
+    padding: 20px;
+}
+
+.breadcrumb {
+    background-color: #c5cee4;
+
+}
+
+.breadcrumb a {
+    text-decoration: none;
+}
+
+.container {
+    max-width: 1000px;
+    padding: 0;
+}
+
+p {
+    margin: 0;
+}
+
+.d-flex a {
+    text-decoration: none;
+    color: #686868;
+}
+
+img {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    object-fit: cover;
+}
+
+.fab.fa-twitter {
+    color: #8ab7f1;
+}
+
+.fab.fa-instagram {
+    color: #E1306C;
+}
+
+.fab.fa-facebook-f {
+    color: #5999ec;
+}
